639: Stop using `org.mozilla.appservices` Gradle plugin.
This sets up "megazord substitutions" just like the Gradle plugin did,
but manually.  The appservices version and megazord definition will
need to be updated when the Android Components version advances, but
at that time you should see build bustage (and not only runtime
bustage).  When that time comes you should be able to clone the
megazord definitions from other consumers like Fenix and/or the
Reference Browser.

The needed dependencies are fetched from the existing
`maven.mozilla.org` repository.

This unwinds #352.